+++ spec/integration/resource/api200/firmware_driver/update_spec.rb
@@ -1,0 +1,6 @@
+require 'spec_helper'
+
+RSpec.describe OneviewSDK::FirmwareDriver, integration: true, type: UPDATE do
+  let(:current_client) { $client }
+  include_examples 'FirmwareDriverUpdateExample', 'integration context'
+end